Switching result last insert id to be long type

+++ src/Rqlite.Client/RqliteCommandResult.cs
@@ -13,12 +13,12 @@ public readonly struct RqliteCommandResult
 	/// <summary>
 	/// For INSERT commands, the ID of the last item inserted.
 	/// </summary>
-	public readonly required int LastInsertId { get; init; }
+	public readonly required long LastInsertId { get; init; }
 
 	/// <summary>
 	/// The number of rows affected by the command.
 	/// </summary>
-	public readonly required int RowsAffected { get; init; }
+	public readonly required long RowsAffected { get; init; }
 
 	internal static RqliteCommandResult Create(ExecuteResponseResult result) =>
 		new() { LastInsertId = result.LastInsertId, RowsAffected = result.RowsAffected };
